refactor: separated pipeline file
