+++ packages/developer-portal/src/actions/apps/index.ts
@@ -7,4 +7,3 @@ export * from './app-revision-detail'
 export * from './app-revision-list'
 export * from './decline-app-revision'
 export * from './create-app-revision'
-export * from './pipeline'
