--- packages/developer-portal/src/actions/apps/pipeline/index.ts
@@ -1,2 +1,0 @@
-export * from './fetch'
-export * from './create'
