+++ packages/developer-portal/src/components/pages/app-detail/__tests__/app-pipeline.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import { AppPipeline } from '../app-pipeline'
+import { AppPipeline } from '../pipeline'
 
 describe('App Pipeline', () => {
   it('should match snapshot', () => {
