+++ packages/developer-portal/src/components/pages/app-detail/app-detail-v8.tsx
@@ -31,7 +31,7 @@ import Routes from '@/constants/routes'
 import { onCloseSubmitAppModal, onShowSubmitAppModal } from '../apps/apps'
 import { SubmitAppWizardModal } from '@/components/ui/submit-app-wizard'
 import ErrorBoundary from '@/components/hocs/error-boundary'
-import { AppPipeline } from './app-pipeline'
+import { AppPipeline } from './pipeline'
 import { MEDIA_INDEX } from '@/constants/media'
 import ImagePlaceHolder from '@/assets/images/default-app-icon.jpg'
 import { cx } from '@linaria/core'
